Refactor framework out of test code
