+++ test/EndToEnd/Framework/AppCollection.cs
@@ -1,4 +1,4 @@
-namespace EndToEndTests.Infrastructure
+namespace EndToEndTests.Framework
 {
     using Xunit;
