+++ test/EndToEnd/Framework/AppFixture.cs
@@ -1,4 +1,4 @@
-namespace EndToEndTests.Infrastructure
+namespace EndToEndTests.Framework
 {
     using Microsoft.Extensions.Logging;
