+++ test/EndToEnd/Tests/TestBase.cs
@@ -1,6 +1,6 @@
-namespace EndToEndTests
+namespace EndToEndTests.Tests
 {
-    using EndToEndTests.Infrastructure;
+    using EndToEndTests.Framework;
     using Xunit;
 
     /// <summary>
